+++ src/brski/brski.cpp
@@ -14,6 +14,7 @@
 extern "C" {
 #include "../utils/log.h"
 #include "../utils/os.h"
+#include "../voucher/serialize.h"
 #include "config.h"
 #include "pledge/pledge_utils.h"
 
@@ -23,13 +24,14 @@ void log_lock_fun(bool lock);
 
 #include "version.h"
 
-const std::string OPT_STRING = ":c:o:dqvh";
+const std::string OPT_STRING = ":c:o:dvh";
 const std::string USAGE_STRING =
-    "\t%s [-c filename] [-o filename] [-d | -q] [-h] [-v] <command>\n";
+    "\t%s [-c filename] [-o filename] [-d] [-h] [-v] <command>\n";
 
 enum class CommandId {
   COMMAND_EXPORT_PVR = 1,
   COMMAND_PLEDGE_REQUEST,
+  COMMAND_SIGN_CERT,
   COMMAND_START_REGISTRAR,
   COMMAND_START_MASA,
 };
@@ -40,12 +42,13 @@ struct command_config {
   const std::string info;
 };
 
-const std::array<struct command_config, 4> command_list = {{
+const std::array<struct command_config, 5> command_list = {{
     {"epvr", CommandId::COMMAND_EXPORT_PVR,
      "\tepvr\t\tExport the pledge voucher request as base64 CMS file"},
     {"preq", CommandId::COMMAND_PLEDGE_REQUEST,
      "\tpreq\t\tSend a pledge-voucher request to the registrar and\n"
      "\t\t\t return the pinned-domain-cert."},
+    {"sign", CommandId::COMMAND_SIGN_CERT, "\tsign\t\tSign a certificate\n"},
     {"registrar", CommandId::COMMAND_START_REGISTRAR,
      "\tregistrar\tStarts the registrar"},
     {"masa", CommandId::COMMAND_START_MASA, "\tmasa\t\tStarts the MASA"},
@@ -85,13 +88,8 @@ static void show_help(const char *name) {
   }
   std::fprintf(stdout, "\nOptions:\n");
   std::fprintf(stdout, "\t-c filename\t Path to the config file\n");
-  std::fprintf(stdout, "\t-o filename\t Path to the exported file\n");
-  std::fprintf(
-      stdout,
-      "\t-d\t\t Verbosity level (use multiple -dd... to increase verbosity)\n");
-  std::fprintf(stdout,
-               "\t-q\t\t Quietness (decreases verbosity) (use twice to hide "
-               "warnings)\n");
+  std::fprintf(stdout, "\t-o filename\t The output file\n");
+  std::fprintf(stdout, "\t-d\t\t Make verbose\n");
   std::fprintf(stdout, "\t-h\t\t Show help\n");
   std::fprintf(stdout, "\t-v\t\t Show app version\n\n");
   std::fprintf(stdout, "Copyright Nquiringminds Ltd\n\n");
@@ -123,7 +121,7 @@ static CommandId get_command_id(const std::string &command_label) {
   throw std::invalid_argument("Unrecognized command: \"" + command_label + '"');
 }
 
-static void process_options(int argc, char *const argv[], int &quietness,
+static void process_options(int argc, char *const argv[], int &verbose,
                             std::string &config_filename,
                             std::string &out_filename, CommandId &command_id) {
   int opt;
@@ -143,10 +141,7 @@ static void process_options(int argc, char *const argv[], int &quietness,
         out_filename.assign(optarg);
         break;
       case 'd':
-        quietness--;
-        break;
-      case 'q':
-        quietness++;
+        verbose = 1;
         break;
       case ':':
         log_cmdline_error("Missing argument for -%c\n", optopt);
@@ -186,65 +181,168 @@ struct BrskiConfig : public brski_config {
   virtual ~BrskiConfig() { free_config_content(this); }
 };
 
+void print_cert(const char *cert, int prefix) {
+  if (prefix)
+    std::fprintf(stdout, "-----BEGIN CERTIFICATE-----\n");
+
+  std::fprintf(stdout, "%s\n", cert);
+
+  if (prefix)
+    std::fprintf(stdout, "-----END CERTIFICATE-----\n");
+}
+
+void print_key(const char *key, int prefix) {
+  if (prefix)
+    std::fprintf(stdout, "-----BEGIN PRIVATE KEY-----\n");
+
+  std::fprintf(stdout, "%s\n", key);
+
+  if (prefix)
+    std::fprintf(stdout, "-----END PRIVATE KEY-----\n");
+}
+
 int main(int argc, char *argv[]) {
-  int quietness = LOGC_INFO;
-  uint8_t log_level = 0;
-  std::string config_filename, out_filename;
+  int verbose = 0;
+  std::string config_filename, in_filename, out_filename;
   CommandId command_id;
+  char outf[255];
 
-  process_options(argc, argv, quietness, config_filename, out_filename,
+  process_options(argc, argv, verbose, config_filename, out_filename,
                   command_id);
-
-  // Clamp quietness to valid log levels enum value
-  // equivalent to C++17 std::clamp(quietness, 0, MAX_LOG_LEVELS - 1)
-  if (quietness >= MAX_LOG_LEVELS) {
-    log_level = MAX_LOG_LEVELS - 1;
-  } else if (quietness < 0) {
-    log_level = 0;
-  } else {
-    log_level = quietness;
-  }
 
   log_set_lock(log_lock_fun);
 
   /* Set the log level */
-  log_set_level(log_level);
+  if (!verbose)
+    log_set_quiet(true);
+  else
+    log_set_level(LOGC_TRACE);
 
   BrskiConfig config;
   if (load_brski_config(config_filename.c_str(), &config) < 0) {
-    std::fprintf(stderr, "load_config fail\n");
+    log_error("load_config fail");
     return EXIT_FAILURE;
   }
 
   struct RegistrarContext *rcontext = NULL;
   struct MasaContext *mcontext = NULL;
+  struct BinaryArray *tls_cert = NULL;
   switch (command_id) {
     case CommandId::COMMAND_EXPORT_PVR:
-      std::fprintf(stdout, "Exporting pledge voucher request to %s",
-                   out_filename.c_str());
-      if (voucher_pledge_request_to_smimefile(&config.pconf,
-                                              config.rconf.tls_cert_path,
-                                              out_filename.c_str()) < 0) {
-        std::fprintf(stderr, "voucher_pledge_request_to_smimefile fail");
+      log_info("Exporting pledge voucher request to %s", out_filename.c_str());
+
+      tls_cert = file_to_x509buf(config.rconf.tls_cert_path);
+      if (tls_cert == NULL) {
+        log_error("file_to_x509buf fail");
         return EXIT_FAILURE;
+      }
+
+      if (out_filename.empty()) {
+        char *base64 =
+            voucher_pledge_request_to_base64(&config.pconf, tls_cert);
+        if (base64 == NULL) {
+          log_error("voucher_pledge_request_to_base64 fail");
+          free_binary_array(tls_cert);
+          return EXIT_FAILURE;
+        }
+        std::fprintf(stdout, "%s\n", base64);
+        sys_free(base64);
+        free_binary_array(tls_cert);
+      } else {
+        snprintf(outf, 255, "%s.smime", out_filename.c_str());
+
+        if (voucher_pledge_request_to_smimefile(&config.pconf, tls_cert, outf) <
+            0) {
+          log_error("voucher_pledge_request_to_smimefile fail");
+          return EXIT_FAILURE;
+        }
+        free_binary_array(tls_cert);
       }
       break;
     case CommandId::COMMAND_PLEDGE_REQUEST: {
-      std::fprintf(stdout, "Pledge voucher request to %s:%d\n",
-                   config.rconf.bind_address, config.rconf.port);
-      std::string response;
+      struct BinaryArray pinned_domain_cert = {};
+
+      log_info("Pledge voucher request to %s:%d", config.rconf.bind_address,
+               config.rconf.port);
       if (post_voucher_pledge_request(&config.pconf, &config.rconf,
-                                      &config.mconf, response) < 0) {
-        std::fprintf(stderr, "post_voucher_pledge_request fail");
+                                      &config.mconf, &pinned_domain_cert) < 0) {
+        log_error("post_voucher_pledge_request fail");
         return EXIT_FAILURE;
       }
-      std::fprintf(stdout, "%s\n", response.c_str());
+
+      if (out_filename.empty()) {
+        char *cert_str = NULL;
+
+        if (serialize_array2base64str(pinned_domain_cert.array,
+                                      pinned_domain_cert.length,
+                                      (uint8_t **)&cert_str) < 0) {
+          log_error("serialize_array2base64str fail");
+          return EXIT_FAILURE;
+        }
+        print_cert(cert_str, 0);
+        sys_free(cert_str);
+      } else {
+        snprintf(outf, 255, "%s.crt", out_filename.c_str());
+        if (certbuf_to_file(&pinned_domain_cert, outf) < 0) {
+          log_error("certbuf_to_file fail");
+          return EXIT_FAILURE;
+        }
+      }
+      free_binary_array_content(&pinned_domain_cert);
       break;
     }
+
+    case CommandId::COMMAND_SIGN_CERT: {
+      log_info("Sign cert %s:%d", config.rconf.bind_address, config.rconf.port);
+      struct BinaryArray out_cert = {};
+      struct BinaryArray out_key = {};
+
+      if (post_sign_cert(&config.pconf, &config.rconf, &config.mconf, &out_cert,
+                         &out_key) < 0) {
+        log_error("post_voucher_pledge_request fail");
+        return EXIT_FAILURE;
+      }
+
+      if (out_filename.empty()) {
+        char *pki_str = NULL;
+
+        if (serialize_array2base64str(out_cert.array, out_cert.length,
+                                      (uint8_t **)&pki_str) < 0) {
+          log_error("serialize_array2base64str fail");
+          return EXIT_FAILURE;
+        }
+        print_cert(pki_str, 0);
+        sys_free(pki_str);
+
+        if (serialize_array2base64str(out_key.array, out_key.length,
+                                      (uint8_t **)&pki_str) < 0) {
+          log_error("serialize_array2base64str fail");
+          return EXIT_FAILURE;
+        }
+        print_cert(pki_str, 0);
+        sys_free(pki_str);
+      } else {
+        snprintf(outf, 255, "%s.crt", out_filename.c_str());
+        if (certbuf_to_file(&out_cert, outf) < 0) {
+          log_error("certbuf_to_file fail");
+          return EXIT_FAILURE;
+        }
+        snprintf(outf, 255, "%s.key", out_filename.c_str());
+        if (keybuf_to_file(&out_key, outf) < 0) {
+          log_error("certbuf_to_file fail");
+          return EXIT_FAILURE;
+        }
+      }
+
+      free_binary_array_content(&out_cert);
+      free_binary_array_content(&out_key);
+      break;
+    }
+
     case CommandId::COMMAND_START_REGISTRAR:
       if (registrar_start(&config.rconf, &config.mconf, &config.pconf,
                           &rcontext) < 0) {
-        std::fprintf(stderr, "https_start fail");
+        log_error("https_start fail");
         return EXIT_FAILURE;
       }
 
@@ -253,7 +351,7 @@ int main(int argc, char *argv[]) {
     case CommandId::COMMAND_START_MASA:
       if (masa_start(&config.rconf, &config.mconf, &config.pconf, &mcontext) <
           0) {
-        std::fprintf(stderr, "https_start fail");
+        log_error("https_start fail");
         return EXIT_FAILURE;
       }