+++ src/brski/pledge/pledge_request.cpp
@@ -35,7 +35,11 @@ extern "C" {
 int post_voucher_pledge_request(struct pledge_config *pconf,
                                 struct registrar_config *rconf,
                                 struct masa_config *mconf,
-                                std::string &response) {
+                                struct BinaryArray *pinned_domain_cert) {
+  int status;
+  struct HttpResponse http_res = {};
+  struct BinaryArray *registrar_tls_cert = NULL;
+
   if (rconf->bind_address == nullptr) {
     log_error("bind_address param is NULL");
     return -1;
@@ -54,7 +58,27 @@ int post_voucher_pledge_request(struct pledge_config *pconf,
   std::string path = PATH_BRSKI_REQUESTVOUCHER;
   std::string content_type = "application/voucher-cms+json";
 
-  char *cms = voucher_pledge_request_to_base64(pconf, rconf->tls_cert_path);
+  /* First run with an empty body to retrieve the certificate */
+  status = https_post_request(pconf->idevid_key_path, pconf->idevid_cert_path,
+                              rconf->bind_address, rconf->port, path, false, "",
+                              content_type, http_res);
+
+  if (status < 0) {
+    log_error("https_post_request fail");
+    return -1;
+  }
+
+  if ((registrar_tls_cert = crypto_cert2buf(http_res.peer_certificate)) ==
+      NULL) {
+    log_error("crypto_cert2buf fail");
+    crypto_free_certcontext(http_res.peer_certificate);
+    return -1;
+  }
+
+  crypto_free_certcontext(http_res.peer_certificate);
+
+  char *cms = voucher_pledge_request_to_base64(pconf, registrar_tls_cert);
+  free_binary_array(registrar_tls_cert);
 
   if (cms == NULL) {
     log_error("voucher_pledge_request_to_base64 fail");
@@ -67,31 +91,37 @@ int post_voucher_pledge_request(struct pledge_config *pconf,
 
   log_info("Request pledge voucher from %s", path.c_str());
 
-  int status = https_post_request(
-      pconf->idevid_key_path, pconf->idevid_cert_path, rconf->bind_address,
-      rconf->port, path, false, body, content_type, response);
+  status = https_post_request(pconf->idevid_key_path, pconf->idevid_cert_path,
+                              rconf->bind_address, rconf->port, path, false,
+                              body, content_type, http_res);
 
   if (status < 0) {
     log_error("https_post_request fail");
     return -1;
   }
+
   if (status >= 400) {
     log_error("post_voucher_pledge_request failed with HTTP code %d and "
               "response: '%s'",
-              status, response.c_str());
+              status, http_res.response.c_str());
+    crypto_free_certcontext(http_res.peer_certificate);
     return -1;
   }
 
-  const char *masa_pledge_voucher_str = response.c_str();
+  if ((registrar_tls_cert = crypto_cert2buf(http_res.peer_certificate)) ==
+      NULL) {
+    log_error("crypto_cert2buf fail");
+    crypto_free_certcontext(http_res.peer_certificate);
+    return -1;
+  }
+
+  crypto_free_certcontext(http_res.peer_certificate);
+
+  const char *masa_pledge_voucher_str = http_res.response.c_str();
   struct BinaryArray masa_pledge_voucher_cms = {};
   struct BinaryArray *nonce = NULL;
-  struct BinaryArray *registrar_tls_cert = NULL;
-  struct BinaryArray *registrar_tls_ca_cert = NULL;
-  struct BinaryArrayList *domain_store = init_array_list();
   struct BinaryArrayList *masa_verify_certs = NULL;
   struct BinaryArrayList *masa_store_certs = NULL;
-  struct BinaryArray pinned_domain_cert = {};
-  char *pinned_domain_cert_base64 = NULL;
 
   int result;
 
@@ -119,23 +149,6 @@ int post_voucher_pledge_request(struct pledge_config *pconf,
     nonce->length = length;
   }
 
-  if ((registrar_tls_cert = file_to_x509buf(rconf->tls_cert_path)) == NULL) {
-    log_error("file_to_x509buf fail");
-    goto post_voucher_pledge_request_fail;
-  }
-
-  if ((registrar_tls_ca_cert = file_to_x509buf(rconf->tls_ca_cert_path)) ==
-      NULL) {
-    log_error("file_to_x509buf fail");
-    goto post_voucher_pledge_request_fail;
-  }
-
-  if (push_array_list(domain_store, registrar_tls_ca_cert->array,
-                      registrar_tls_ca_cert->length, 0) < 0) {
-    log_error("push_array_list");
-    goto post_voucher_pledge_request_fail;
-  }
-
   if (load_cert_files(pconf->cms_verify_certs_paths, &masa_verify_certs) < 0) {
     log_error("load_cert_files");
     goto post_voucher_pledge_request_fail;
@@ -148,43 +161,169 @@ int post_voucher_pledge_request(struct pledge_config *pconf,
 
   result = verify_masa_pledge_voucher(
       &masa_pledge_voucher_cms, pconf->serial_number, nonce, registrar_tls_cert,
-      domain_store, masa_verify_certs, masa_store_certs, NULL,
-      &pinned_domain_cert);
+      NULL, masa_verify_certs, masa_store_certs, NULL, pinned_domain_cert);
 
   if (result < 0) {
     log_error("verify_masa_pledge_voucher fail");
     goto post_voucher_pledge_request_fail;
   }
 
-  if (serialize_array2base64str(pinned_domain_cert.array,
-                                pinned_domain_cert.length,
-                                (uint8_t **)&pinned_domain_cert_base64) < 0) {
-    log_error("serialize_array2base64str fail");
-    goto post_voucher_pledge_request_fail;
-  }
-
-  response.assign(pinned_domain_cert_base64);
-
   free_binary_array_content(&masa_pledge_voucher_cms);
   free_binary_array(nonce);
   free_binary_array(registrar_tls_cert);
-  free_binary_array(registrar_tls_ca_cert);
-  free_array_list(domain_store);
   free_array_list(masa_verify_certs);
   free_array_list(masa_store_certs);
-  free_binary_array_content(&pinned_domain_cert);
-  free(pinned_domain_cert_base64);
   return 0;
 
 post_voucher_pledge_request_fail:
   free_binary_array_content(&masa_pledge_voucher_cms);
   free_binary_array(nonce);
   free_binary_array(registrar_tls_cert);
-  free_binary_array(registrar_tls_ca_cert);
-  free_array_list(domain_store);
   free_array_list(masa_verify_certs);
   free_array_list(masa_store_certs);
+  return -1;
+}
+
+std::string create_cert_string(const char *cert) {
+  std::string out = "-----BEGIN CERTIFICATE-----\n";
+  out += std::string(cert) + "\n";
+  out += "-----END CERTIFICATE-----";
+
+  return out;
+}
+
+int generate_sign_cert(struct BinaryArray *scert_cert,
+                       struct BinaryArray *scert_key) {
+  struct crypto_cert_meta sign_cert_meta = {
+      .serial_number = 12345,
+      .not_before = 0,
+      // Long-lived pledge certificate
+      .not_after_absolute = (char *)"99991231235959Z",
+      .issuer = NULL,
+      .subject = NULL,
+      .basic_constraints = (char *)"CA:false"};
+
+  if ((sign_cert_meta.issuer = init_keyvalue_list()) == NULL) {
+    log_error("init_keyvalue_list fail");
+    return -1;
+  }
+
+  if ((sign_cert_meta.subject = init_keyvalue_list()) == NULL) {
+    log_error("init_keyvalue_list fail");
+    free_keyvalue_list(sign_cert_meta.issuer);
+    return -1;
+  }
+
+  if (push_keyvalue_list(sign_cert_meta.subject, (char *)"C", (char *)"IE") <
+      0) {
+    log_error("push_keyvalue_list fail");
+    goto generate_sign_cert_err;
+  }
+
+  if (push_keyvalue_list(sign_cert_meta.subject, (char *)"CN",
+                         (char *)"pledge-domain-cert") < 0) {
+    log_error("push_keyvalue_list fail");
+    goto generate_sign_cert_err;
+  }
+
+  if ((scert_key->length = (size_t)crypto_generate_eckey(&scert_key->array)) <
+      0) {
+    log_error("crypto_generate_eckey fail");
+    goto generate_sign_cert_err;
+  }
+
+  if ((scert_cert->length = (size_t)crypto_generate_eccert(
+           &sign_cert_meta, scert_key->array, scert_key->length,
+           &scert_cert->array)) < 0) {
+    free_binary_array_content(scert_key);
+    goto generate_sign_cert_err;
+  }
+
+  free_keyvalue_list(sign_cert_meta.issuer);
+  free_keyvalue_list(sign_cert_meta.subject);
+
+  return 0;
+
+generate_sign_cert_err:
+  free_keyvalue_list(sign_cert_meta.issuer);
+  free_keyvalue_list(sign_cert_meta.subject);
+
+  return -1;
+}
+
+int post_sign_cert(struct pledge_config *pconf, struct registrar_config *rconf,
+                   struct masa_config *mconf, struct BinaryArray *out_cert,
+                   struct BinaryArray *out_key) {
+  std::string pinned_cert, response, ca, body;
+  struct BinaryArray pinned_domain_cert = {};
+  int status;
+  char *pki_str = NULL;
+  std::string path = PATH_BRSKI_SIGNCERT;
+  std::string content_type = "application/voucher-cms+json";
+  std::string registrar_ca_cert;
+  ssize_t length;
+
+  if (generate_sign_cert(out_cert, out_key) < 0) {
+    log_error("generate_sign_cert");
+    return -1;
+  }
+
+  if (post_voucher_pledge_request(pconf, rconf, mconf, &pinned_domain_cert) <
+      0) {
+    log_error("post_voucher_pledge_request fail");
+    goto post_sign_cert_err;
+  }
+
+  if (serialize_array2base64str(pinned_domain_cert.array,
+                                pinned_domain_cert.length,
+                                (uint8_t **)&pki_str) < 0) {
+    log_error("serialize_array2base64str fail");
+    goto post_sign_cert_err;
+  }
+
+  registrar_ca_cert = create_cert_string(pki_str);
+  sys_free(pki_str);
+
+  if (serialize_array2base64str(out_cert->array, out_cert->length,
+                                (uint8_t **)&pki_str) < 0) {
+    log_error("serialize_array2base64str fail");
+    goto post_sign_cert_err;
+  }
+  body = pki_str;
+  sys_free(pki_str);
+
+  status = https_post_request_ca(
+      pconf->idevid_key_path, pconf->idevid_cert_path, registrar_ca_cert,
+      rconf->bind_address, rconf->port, path, body, content_type, response);
+
+  if (status < 0) {
+    log_error("https_post_request fail");
+    goto post_sign_cert_err;
+  }
+
+  if (status >= 400) {
+    log_error("post_voucher_pledge_request_ca failed with HTTP code %d and "
+              "response: '%s'",
+              status, response.c_str());
+    goto post_sign_cert_err;
+  }
+
+  free_binary_array_content(out_cert);
+  pki_str = (char *)response.c_str();
+
+  if ((length = serialize_base64str2array(
+           (const uint8_t *)pki_str, strlen(pki_str), &out_cert->array)) < 0) {
+    log_errno("serialize_base64str2array fail");
+    goto post_sign_cert_err;
+  }
+  out_cert->length = length;
+
   free_binary_array_content(&pinned_domain_cert);
-  free(pinned_domain_cert_base64);
+  return 0;
+
+post_sign_cert_err:
+  free_binary_array_content(out_cert);
+  free_binary_array_content(out_key);
+  free_binary_array_content(&pinned_domain_cert);
   return -1;
 }