+++ src/brski/registrar/registrar_api.hpp
@@ -66,4 +66,19 @@ int registrar_enrollstatus(const RequestHeader &request_header,
                            ResponseHeader &response_header,
                            std::string &response, void *context);
 
+/**
+ * @brief Registrar sign certificate
+ * Not part of the specification.
+ *
+ * @return The HTTP status code.
+ * @retval 200 OK.
+ * @retval 400 Bad Request (malformed request).
+ * @retval 500 Internal Server Error.
+ * @retval 502 Bad Gateway
+ */
+int registrar_signcert(const RequestHeader &request_header,
+                       const std::string &request_body,
+                       CRYPTO_CERT peer_certificate,
+                       ResponseHeader &response_header, std::string &response,
+                       void *context);
 #endif