+++ src/voucher/crypto.h
@@ -82,6 +82,24 @@ enum CRYPTO_CERTIFICATE_TYPE {
 };
 
 /**
+ * @brief Makes a copy of the certificate structire
+ *
+ * @param[in] cert The input certificate structure
+ * @return CRYPTO_CERT certificate context, NULL on failure
+ */
+CRYPTO_CERT crypto_copycert(CRYPTO_CERT cert);
+
+/**
+ * @brief Convert a cert context to a DER binary array
+ *
+ * Caller is responsible for freeing the binary array
+ *
+ * @param[in] cert The input certificate structure
+ * @return struct BinaryArray * the output DER binary array, NULL on failure
+ */
+struct BinaryArray *crypto_cert2buf(CRYPTO_CERT cert);
+
+/**
  * @brief Parses a certificate
  *
  * @param[in] cert The input certificate structure
@@ -131,6 +149,13 @@ ssize_t crypto_generate_eckey(uint8_t **key);
  * @param[in] ctx The key context
  */
 void crypto_free_keycontext(CRYPTO_KEY ctx);
+
+/**
+ * @brief Frees a certificate context
+ *
+ * @param[in] ctx The certificate context
+ */
+void crypto_free_certcontext(CRYPTO_KEY cert);
 
 #if __GNUC__ >= 11 // this syntax will throw an error in GCC 10 or Clang
 #define __must_crypto_free_keycontext                                          \